Created types folder
