+++ ilp-api/utils/ilp-utils/wallet.ts
@@ -1,20 +1,6 @@
 import { client } from "./client";
 import { OpenPaymentsClientError } from "@interledger/open-payments";
-
-export interface IWalletAddressResponse {
-  id: string;
-  publicName: string;
-  assetCode: string;
-  assetScale: number;
-  authServer: string;
-  resourceServer: string;
-}
-
-export interface InvalidWalletResponse extends OpenPaymentsClientError {
-  status: number;
-  description: string;
-}
-
+import { InvalidWalletResponse, IWalletAddressResponse } from "../types/wallet";
 const validateWalletAddress = async (
   walletAddress: string
 ): Promise<IWalletAddressResponse> => {
